Resolved Integrate version 2.0.0 BasicServices and ApplicationPattern Fixes #227
