+++ server/package.json
@@ -21,6 +21,8 @@
     "node-statsd": "^0.1.1",
     "oas3-tools": "^2.2.3",
     "randexp": "^0.5.3",
-    "response-time": "^2.3.2"
+    "response-time": "^2.3.2",
+    "onf-core-model-ap-bs": "2.0.0-alpha.3", 
+    "onf-core-model-ap": "2.0.0-alpha.3"
   }
 }
